requirements.txt
pygame
mutagen
pdfminer.six